name'ing update

+++ app/Http/Controllers/API/BookmarkController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers\API;
 use App\Bookmark;
 use App\Factories\BookmarkFactory;
 use App\Http\Services\BookmarksService;
+use App\Post;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Auth;
+use Symfony\Component\HttpFoundation\Response;
 
 class BookmarkController extends BaseController
 {
@@ -21,17 +23,23 @@ class BookmarkController extends BaseController
 
     public function bookmark(int $post): JsonResponse
     {
-        if ($this->bookmarksService->isBookmarked($post)) {
-            Bookmark::where('user_id', Auth::id())->where('post', $post)->delete();
-        } else {
-            $bookmark = $this->bookmarkFactory->create([
-                'post' => $post,
-                'user_id' => Auth::id()
-            ], Bookmark::class);
-
-            $bookmark->save();
+        if (!Post::find($post)) {
+            return $this->return(['error' => 'This post does not exist!'], Response::HTTP_NOT_FOUND);
         }
 
-        return $this->return();
+        if ($this->bookmarksService->isBookmarked($post)) {
+            Bookmark::where('user_id', Auth::id())->where('post', $post)->delete();
+
+            return $this->return(responseCode: Response::HTTP_NO_CONTENT);
+        }
+
+        $bookmark = $this->bookmarkFactory->create([
+            'post' => $post,
+            'user_id' => Auth::id()
+        ], Bookmark::class);
+
+        $bookmark->save();
+
+        return $this->return(compact('bookmark'), Response::HTTP_CREATED);
     }
 }
